requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-allauth==0.44.0
future==0.18.2
idna==2.10
mailchimp-marketing==3.0.43
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyVimeo==1.1.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
stripe==2.57.0
tinydb==4.4.0
tuspy==0.2.4
urllib3==1.26.4
